requirements.txt
flask==3.0.0
pycryptodome==3.20.0